isspace() and tolower() are int->int as per the specs. These are generally implemented as table lookups and the compiler gives an error about a character being used as an table index
git-svn-id: https://dev.csoft.net/agar/trunk@9442 86a84d31-9702-0410-8f1e-a60cab7bd424

+++ gui/file_dlg.c
@@ -431,7 +431,7 @@ CheckAccessAndChoose(AG_FileDlg *fd)
 	char *s;
 
 	for (s = &fd->cfile[0]; *s != '\0'; s++) {
-		if (!isspace(*s))
+		if (!isspace((int) *s))
 			break;
 	}
 	if (*s == '\0')
